⚡️ optimize image quality and dimensions in multiple components
