+++ app/components/carousel/Carousel.tsx
@@ -1,122 +1,17 @@
 "use client";
 import type { ItemDetails } from "@/app/types/getDetails";
-import { Button } from "@/components/ui/button";
 import type { EmblaCarouselType } from "embla-carousel";
 import useEmblaCarousel from "embla-carousel-react";
-import {
-	BookOpen,
-	ChevronLeft,
-	ChevronRight,
-	InfoIcon,
-	PlayIcon,
-	Volume2,
-	VolumeX,
-} from "lucide-react";
-import Image from "next/image";
-import Link from "next/link";
-import {
-	createContext,
-	useContext,
-	useEffect,
-	useMemo,
-	useRef,
-	useState,
-} from "react";
+import { ChevronLeft, ChevronRight } from "lucide-react";
+import { createContext, useContext, useEffect, useState } from "react";
+import AnimeDetailComponent from "./anime/animeDetailComponent";
+import MangaDetailComponent from "./manga/mangaDetailComponent";
 
 interface EmblaCarouselProps {
 	Details: ItemDetails[];
 }
 const language = process.env.DEFAULT_LANGUAGE;
 const data = require(`@/locales/${language}.json`);
-function MangaDetailComponent({
-	detail,
-}: {
-	detail: ItemDetails;
-}) {
-	const imageSrc = useMemo(
-		() => `/api/image?path=${detail.name}/manga/01/01-001.webp`,
-		[detail.name],
-	);
-
-	return (
-		<div
-			className="relative flex-shrink-0 w-full h-126 flex flex-col items-center"
-			key={detail.name}
-		>
-			<Image
-				src={imageSrc}
-				alt={"cover image back"}
-				className="object-cover opacity-25 blur-lg"
-				fill
-				sizes="10vw"
-				quality={1}
-				placeholder="blur"
-				blurDataURL="data:image/gif;base64,R0lGODlhAQABAIAAAAUEBAAAACwAAAAAAQABAAACAkQBADs="
-			/>
-			<div className="relative w-full flex justify-center mt-8">
-				<div className="relative w-full flex justify-center">
-					<div className="relative w-56 h-80">
-						<Image
-							src={imageSrc}
-							alt=""
-							className="object-scale-down md:transform  md:transition-transform md:duration-500 md:rotate-12 md:hover:rotate-0 md:hover:scale-110"
-							quality={50}
-							fill
-							placeholder="blur"
-							blurDataURL="data:image/gif;base64,R0lGODlhAQABAIAAAAUEBAAAACwAAAAAAQABAAACAkQBADs="
-						/>
-					</div>
-				</div>
-			</div>
-
-			<div className="relative w-full h-full flex items-end mb-4">
-				<div className="relative w-32 m-2">
-					<Image
-						src={imageSrc}
-						alt={"cover image front"}
-						className="object-contain"
-						quality={50}
-						width={300}
-						height={450}
-						sizes="(min-width: 1080px) 216px, (min-width: 1000px) calc(-15vw + 374px), (min-width: 780px) calc(15.5vw + 72px), (min-width: 560px) 224px, (min-width: 380px) calc(18.75vw + 123px), calc(75vw - 83px)"
-						placeholder="blur"
-						blurDataURL="data:image/gif;base64,R0lGODlhAQABAIAAAAUEBAAAACwAAAAAAQABAAACAkQBADs="
-					/>
-				</div>
-
-				<div className="w-1/2 text-left ml-4 space-y-2 flex flex-col justify-center">
-					<h1 className="text-xl">{detail.name}</h1>
-					<p className="text-xs pr-2 line-clamp-2 lg:line-clamp-3  max-w-lg ">
-						{detail.synopsis}
-					</p>
-					<p className="text-sm font-normal">
-						{detail.volumes?.length ?? 0}{" "}
-						{`${(detail.volumes?.[0]?.type ?? data.carousel.volumes).trim()}s`}
-					</p>
-					<div>
-						<Link href={`/manga/${detail.name}/1`}>
-							<Button>
-								<BookOpen className="mr-2" />
-								<span className="sm:hidden">{data.carousel.read}</span>
-								<span className="hidden sm:block">{data.carousel.start}</span>
-							</Button>
-						</Link>
-
-						<Link href={`/detail/${detail.name}`}>
-							<Button
-								variant="ghost"
-								className="rounded-full p-2 m-4 bg-opacity-50 dark:bg-black bg-white"
-								aria-label="More Info"
-							>
-								<InfoIcon />
-							</Button>
-						</Link>
-					</div>
-				</div>
-			</div>
-		</div>
-	);
-}
 
 // MuteContext
 const MuteContext = createContext<
@@ -143,206 +38,36 @@ export const useMute = () => {
 	return context;
 };
 
-function AnimeDetailComponent({
-	detail,
-	isActive,
-	emblaApi,
-}: {
-	detail: ItemDetails;
-	isActive: boolean;
-	emblaApi: EmblaCarouselType | null;
-}) {
-	const videoRef = useRef<HTMLVideoElement>(null);
-	const [isMuted, setIsMuted] = useMute();
-
-	useEffect(() => {
-		if (isActive && videoRef.current) {
-			videoRef.current.play();
-		} else if (videoRef.current) {
-			videoRef.current.pause();
-			videoRef.current.currentTime = 0;
-		}
-	}, [isActive]);
-
-	const handleMute = () => {
-		if (videoRef.current) {
-			videoRef.current.muted = !videoRef.current.muted;
-			setIsMuted(!isMuted);
-		}
-	};
-
-	const videoSrc = useMemo(
-		() => `/api/video?type=preview&videoId=${detail.name}`,
-		[detail.name],
-	);
-	const thumbnailSrc = useMemo(
-		() => `/api/image?path=${detail.name}/anime/thumbnail.webp`,
-		[detail.name],
-	);
-
-	const posterSrc = useMemo(
-		() =>
-			`/api/image?type=thumbnail&path=${detail.name}/anime/Season01/01-001.webp`,
-		[detail.name],
-	);
-
-	const [progress, setProgress] = useState(0);
-
-	useEffect(() => {
-		const interval = setInterval(() => {
-			if (videoRef.current) {
-				const progress =
-					(videoRef.current.currentTime / videoRef.current.duration) * 100;
-				setProgress(progress);
-			}
-		}, 1000);
-
-		return () => {
-			clearInterval(interval);
-		};
-	}, []);
-
-	const [isIntersecting, setIsIntersecting] = useState(false);
-	const ref = useRef<HTMLDivElement>(null);
-
-	useEffect(() => {
-		const observer = new IntersectionObserver(
-			([entry]) => {
-				setIsIntersecting(entry.isIntersecting);
-				if (entry.isIntersecting && videoRef.current && isActive) {
-					videoRef.current.play();
-				}
-			},
-			{
-				root: null,
-				rootMargin: "0px",
-				threshold: 0.1,
-			},
-		);
-
-		if (ref.current) {
-			const currentRef = ref.current;
-			observer.observe(currentRef);
-
-			return () => {
-				observer.unobserve(currentRef);
-			};
-		}
-	}, [isActive]);
-
-	return (
-		<div
-			className="relative flex-shrink-0 w-full h-126"
-			key={detail.name}
-			ref={ref}
-		>
-			{isIntersecting && (
-				<>
-					<video
-						poster={posterSrc}
-						ref={videoRef}
-						src={videoSrc}
-						muted={isMuted}
-						className="absolute w-full h-full object-cover"
-						onEnded={() => emblaApi?.scrollNext()}
-						autoPlay
-						playsInline
-						preload="none"
-					/>
-
-					<div className="relative w-full h-full flex items-end">
-						<div className="relative w-32 m-2">
-							<Image
-								src={thumbnailSrc}
-								alt={"cover image front"}
-								className="object-contain"
-								quality={50}
-								width={300}
-								height={450}
-								sizes="(min-width: 1080px) 216px, (min-width: 1000px) calc(-15vw + 374px), (min-width: 780px) calc(15.5vw + 72px), (min-width: 560px) 224px, (min-width: 380px) calc(18.75vw + 123px), calc(75vw - 83px)"
-								placeholder="blur"
-								blurDataURL="data:image/gif;base64,R0lGODlhAQABAIAAAAUEBAAAACwAAAAAAQABAAACAkQBADs="
-								priority={true}
-							/>
-						</div>
-						<div className="w-1/2 text-left ml-4 space-y-2 flex flex-col justify-center">
-							<h1 className="text-xl text-white">{detail.name}</h1>
-							<p className="text-xs pr-2 line-clamp-2 lg:line-clamp-3 max-w-lg text-white">
-								{detail.synopsis}
-							</p>
-							<p className="text-sm font-normal text-white">
-								{detail.seasons?.length ?? 0} {data.carousel.seasons},{" "}
-								{detail.episodeNumber} {data.carousel.episodes}
-							</p>
-							<div>
-								<Link href={`/anime/${detail.name}/season01/episode01`}>
-									<Button>
-										<PlayIcon className="mr-2" />
-										<span className="sm:hidden">{data.carousel.watch}</span>
-										<span className="hidden sm:block">
-											{data.carousel.startW}
-										</span>
-									</Button>
-								</Link>
-
-								<Link href={`/detail/${detail.name}`}>
-									<Button
-										variant="ghost"
-										className="rounded-full p-2 m-4 bg-opacity-50 dark:bg-black bg-white"
-										aria-label="More Info"
-									>
-										<InfoIcon />
-									</Button>
-								</Link>
-								<Button
-									variant="ghost"
-									onClick={handleMute}
-									className="absolute right-1 bottom-8 md:bottom-1/4 transform translate-y-1/2 m-4 dark:bg-black bg-white bg-opacity-50 rounded-full p-1"
-									aria-label="Mute/Unmute Video"
-								>
-									{isMuted ? (
-										<VolumeX className="w-10 h-10" />
-									) : (
-										<Volume2 className="w-10 h-10" />
-									)}
-								</Button>
-							</div>
-						</div>
-					</div>
-					<div className="h-1 dark:bg-black bg-white w-full bottom-0 absolute">
-						<div
-							style={{ width: `${progress}%` }}
-							className="absolute h-1 bg-red-600"
-						/>
-					</div>
-				</>
-			)}
-		</div>
-	);
-}
-
 function DetailComponent({
 	detail,
 	isActive,
 	emblaApi,
+	isPriority,
 }: {
 	detail: ItemDetails;
 	isActive: boolean;
 	emblaApi: EmblaCarouselType | null;
+	isPriority: boolean;
 }) {
 	if (detail.types.includes("anime")) {
-		const animeDetail = detail;
 		return (
 			<AnimeDetailComponent
-				detail={animeDetail}
+				detail={detail}
 				isActive={isActive}
 				emblaApi={emblaApi}
+				data={data}
+				isPriority={isPriority}
 			/>
 		);
 	}
 	if (detail.types.includes("manga")) {
-		const mangaDetail = detail;
-		return <MangaDetailComponent detail={mangaDetail} />;
+		return (
+			<MangaDetailComponent
+				detail={detail}
+				data={data}
+				isPriority={isPriority}
+			/>
+		);
 	}
 	return null;
 }
@@ -378,12 +103,14 @@ export default function EmblaCarousel(props: EmblaCarouselProps) {
 				<div className="flex">
 					{Details.map((Detail, index) => {
 						const isActive = index === activeIndex;
+						const isPriority = index === 0;
 						return (
 							<DetailComponent
 								detail={Detail}
 								isActive={isActive}
 								key={Detail.name}
 								emblaApi={emblaApi || null}
+								isPriority={isPriority}
 							/>
 						);
 					})}
